FIX 게시글 컨트롤러 GET "/post" page 파라미터 Long => Integer

+++ src/main/java/com/spring/clone/post/PostController.java
@@ -18,7 +18,7 @@ public class PostController {
 
     @GetMapping("/post")
     public Map<String, Object> getPostsOrderByCreatedAtDesc(
-            @RequestParam("page") Long page
+            @RequestParam("page") Integer page
     ) {
         Map<String, Object> result = new HashMap<>();
         List<Post> posts = postService.getPostsOrderByCreatedAtDesc(page);
